# doc/api.yaml
openapi: 3.0.0
info:
  title: Chat Application API
  description: Chat application REST API documentation
  version: 1.0.0
  contact:
    name: Chat App Team
    email: support@chatapp.com
    url: https://chatapp.com/support
servers:
- url: /api/v1
  description: API foundation path
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT authentication using the Bearer scheme.
        Example: "Authorization: Bearer {token}"
  responses:
    BadRequest:
      description: >
        The request was malformed or contained invalid parameters.
        Check the error details for specific validation issues.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 400
            message: "Invalid request parameters"
            trace: "req-123456"
            error: "validation.failed"
            errors:
              - "username: Must be between 3-50 characters"
              - "email: Invalid email format"
...
    Forbidden:
      description: >
        Authenticated user does not have permission to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    ConversationListEnvelope:
      description: Envelope with a list of conversation summaries for the current user.
      allOf:
        - $ref: '#/components/schemas/BaseSuccessResponse'
        - type: object
          required: 
            - items
          properties:
            items:
              type: array
              description: List of conversation resources
              minItems: 0
              maxItems: 100
              items:
                $ref: '#/components/schemas/Conversation'
...
paths:
...
  /conversations:
    post:
      tags:
        - conversations
      operationId: startConversation
      summary: Start a new conversation
      description: Start a new group conversation with selected members
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Request to start a new conversation
              required:
                - member_ids
                - name
              properties:
                member_ids:
                  type: array
                  description: List of member IDs to include in the conversation
                  items:
                    type: string
                    pattern: ^[a-zA-Z0-9-]+$
                    description: User ID of a conversation member
                    minLength: 1
                    maxLength: 64
                  minItems: 1
                  maxItems: 100
                name:
                  type: string
                  description: Name for the new conversation
                  minLength: 1
                  maxLength: 100
                  pattern: ^[a-zA-Z0-9\s'-]+$
      responses:
        '201':
          description: Conversation started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationListEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    get:
      tags:
      - conversations
      operationId: getMyConversations
      summary: Get my conversations list
      description: Retrieve list of user's conversations with last message preview
      security:
      - BearerAuth: []
      responses:
        '200':
          description: Conversations list retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationListEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /session:
    post:
      tags:
      - auth
      operationId: doLogin
      summary: Simplified user login
      description: Login with just username (creates account if new)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Login request
              required:
              - name
              properties:
                name:
                  type: string
                  description: User's display name
                  minLength: 3
                  maxLength: 16
                  pattern: ^[a-zA-Z0-9\s'-]+$
                  example: Maria
      responses:
        '201':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthEnvelope'
              example:
                code: 201
                message: Login successful
                data:
                  user:
                    id: usr-999
                    username: maria
                    email: maria@example.com
                    avatar_url: https://example.com/avatars/maria.jpg
                    gender: unspecified
                  token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJ1c3ItOTk5IiwiaWF0IjoxNjk2MDAwMDB9.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /register:
    post:
      tags:
      - auth
      operationId: doRegister
      summary: Register a new user
      description: Create a new user account with the provided information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthEnvelope'
              example:
                code: 201
                message: User registered successfully
                data:
                  user:
                    id: usr-123456
                    username: john_doe
                    email: john@example.com
                    avatar_url: https://example.com/avatars/john_doe.jpg
                    gender: male
                  token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJ1c3ItOTk5IiwiaWF0IjoxNjk2MDAwMDB9.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/set_username:
    put:
      tags:
      - users
      operationId: setMyUserName
      summary: Change username
      description: Update the current user's username
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Request to update username
              properties:
                username:
                  type: string
                  description: New username
                  minLength: 1
                  maxLength: 50
                  pattern: ^[a-zA-Z0-9_-]+$
      responses:
        '200':
          description: Username updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseSuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/set_photo:
    put:
      tags:
      - users
      operationId: setMyPhoto
      summary: Change profile photo
      description: Update the current user's profile picture
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
              - upload
              description: Request to update profile photo
              properties:
                upload:
                  type: string
                  format: binary
                  description:  Image file (JPEG/PNG)
                  minLength: 1
                  maxLength: 10485760
      responses:
        '200':
          description: Profile photo updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileUploadEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/me:
    get:
      tags:
      - users
      operationId: getUserInfo
      summary: Get current user info
      description: Retrieve current user's profile information
      security:
      - BearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEnvelope'
              example:
                code: 200
                message: User information retrieved
                data:
                  user:
                    id: usr-123456
                    username: john_doe
                    name: John Doe
                    email: john@example.com
                    avatar_url: https://example.com/avatars/john_doe.jpg
                    gender: male
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/search:
    get:
      tags:
      - users
      summary: Search users by keyword
      description: Search users by partial name, username or email
      operationId: searchUsers
      parameters:
      - name: q
        in: query
        required: true
        description: Search query
        schema:
          type: string
          minLength: 1
          maxLength: 255
          pattern: ^[a-zA-Z0-9\s@.-]+$
          description: Search keyword to match against user names,usernames, or email addresses
      security:
      - BearerAuth: []
      responses:
        '200':
          description: Matching users found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEnvelopeCollection'
              example:
                code: 200
                message: Matching users found
                items:
                  - id: usr-1
                    username: alice
                    avatar_url: https://example.com/avatars/alice.jpg
                    email: alice@example.com
                    gender: female
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/profile/{user_id}:
    parameters:
      - $ref: '#/components/parameters/user_id'
    get:
      tags:
      - users
      operationId: getUserProfile
      summary: Get user profile
      description: Retrieve detailed profile information for a specific user
      security:
      - BearerAuth: []
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEnvelope'
              example:
                code: 200
                message: User information retrieved
                data:
                  user:
                    id: usr-123456
                    username: john_doe
                    name: John Doe
                    email: john@example.com
                    avatar_url: https://example.com/avatars/john_doe.jpg
                    gender: male
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /groups:
    post:
      tags:
      - groups
      operationId: createGroup
      summary: Create group
      description: Create group and associated conversation
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Create a new group 
              required: 
              - name
              - member_ids
              properties:
                name:
                  type: string
                  description: Group display name 
                  minLength: 1
                  maxLength: 100
                  pattern: "^[a-zA-Z0-9\\s'-]+$"
                member_ids:
                  type: array
                  description: IDs of users to include
                  minItems: 1
                  maxItems: 100
                  items:
                    type: string
                    description: User ID to add to the group
                    pattern: '^[a-zA-Z0-9-]+$'
                    minLength: 1
                    maxLength: 64
      responses:
        '201':
          description: Group created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupEnvelope'
              example:
                code: 201
                message: Group created
                data:
                  group:
                    id: '1'
                    name: Group 1
                    conversation_id: 'c-1'  
                    members:
                      - user_id: '1'
                        role: admin
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    get:
      tags:
      - groups
      operationId: getGroupsList
      summary: Get groups list
      description: Retrieve list of user's groups
      security:
      - BearerAuth: []
      responses:
        '200':
          description: Groups list retrieved successfully
          content:
            application/json:
              schema:
                 $ref: '#/components/schemas/GroupEnvelopeCollection'
              example:
                code: 200
                message: Groups list retrieved
                items:                        
                  - id: '1'
                    name: Group 1
                    conversation_id: 'c-1'        
                    members:
                      - user_id: '1'
                        role: admin
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /groups/{id}:
    parameters:
      - $ref: '#/components/parameters/id'
    get:
      tags:
      - groups
      operationId: getGroupDetail
      summary: Get group details
      description: Retrieve details of a specific group
      security:
      - BearerAuth: []
      responses:
        '200':
          description: Group details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupEnvelope'
              example:
                code: 200
                message: Group details retrieved
                data:
                    group:
                      id: '1'
                      name: Group 1
                      conversation_id: 'c-1' 
                      members:
                        - user_id: '1'
                          role: admin
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /groups/{id}/name:
    parameters:
      - $ref: '#/components/parameters/id'
    put:
      tags:
      - groups
      operationId: setGroupName
      summary: Update group name
      description: Update the name of an existing group
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGroupNameRequest'
      responses:
        '200':
          description: Group name updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseSuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /groups/{id}/photo:
    parameters:
      - $ref: '#/components/parameters/id'
    put:
      tags:
      - groups
      operationId: setGroupPhoto
      summary: Change group photo
      description: Update the group's profile picture
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              description: Request to update group photo
              required:
                - upload
              properties:
                upload:
                  type: string
                  format: binary
                  description: Group photo file
                  minLength: 1
                  maxLength: 10485760
      responses:
        '200':
          description: Group photo updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileUploadEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /groups/{id}/members:
    parameters:
      - $ref: '#/components/parameters/id'
    post:
      tags:
      - groups
      operationId: addToGroup
      summary: Add group members
      description: Add new members to an existing group
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddGroupMembersRequest'
      responses:
        '200':
          description: Members added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseSuccessResponse'
              example:
                code: 200
                message: Members added
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
      - groups
      operationId: leaveGroup
      summary: Leave group
      description: Current user leaves the specified group
      security:
      - BearerAuth: []
      responses:
        '200':
          description: Successfully left the group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseSuccessResponse'
              example:
                code: 200
                message: Left the group
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /messages:
    get:
      tags:
        - messages
      operationId: getConversation
      summary: Get conversation messages
      description: Retrieve messages from a specific conversation
      security:
        - BearerAuth: []
      parameters:
        - name: conversation_id
          in: query
          required: true
          description: ID of the conversation
          schema:
            type: string
            pattern: ^[a-zA-Z0-9-]+$
            description: Unique identifier of the conversation to retrieve messages from
            minLength: 1
            maxLength: 64
        - name: limit
          in: query
          required: false
          description: Maximum number of messages to return
          schema:
            type: integer
            format: int64
            minimum: 1
            maximum: 100
            default: 20
            description: Maximum number of messages to return in a single page
        - name: before
          in: query
          description: Get messages before this timestamp (ISO 8601 format)
          schema:
            type: string
            description: ISO 8601 timestamp to get messages before this time
            format: date-time
            minLength: 20
            maxLength: 35
      responses:
        '200':
          description: Messages retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResourceEnvelopeCollection'
              example:
                code: 200
                message: Messages retrieved successfully
                data:
                  messages:
                    - id: '1'
                      content: Hello!
                      sender_id: '1'
                      conversation_id: 'c-1'
                      created_at: '2023-10-01T12:00:00Z'
                      type: text
                      status: delivered
                  pagination:
                    total: 120
                    limit: 20
                    offset: 0
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    
    post:
      tags:
      - messages
      operationId: sendMessage
      summary: Send a message
      description: Send a new message to a user or group
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Send a message
              required:
              - conversation_id
              - content
              properties:
                conversation_id:
                  type: string
                  description: ID of the target conversation
                  pattern: ^[a-zA-Z0-9-]+$
                  minLength: 1 
                  maxLength: 64
                content:
                  type: string
                  description: Message content
                  minLength: 1
                  maxLength: 1000
                  pattern: '^[\s\S]{1,1000}$'
                type:
                  type: string
                  description: Message type
                  enum: 
                  - text
                  - image
                  - video
                  - file
                  default: text
      responses:
        '201':
          description: Message sent successfully
          content:
            application/json:
              schema:
               $ref: '#/components/schemas/MessageResourceEnvelope'
              example:
                code: 201
                message: Message sent successfully
                data:
                  resource:
                    id: '123'
                    content: Hello!
                    sender_id: '1'
                    conversation_id: 'c-1'
                    created_at: '2023-10-01T12:00:00Z'
                    type: text
                    status: sent
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /messages/{id}:
    parameters:
      - $ref: '#/components/parameters/id'
    get:
      tags:
      - messages
      operationId: getMessageById
      summary: Get a specific message
      description: Retrieve details of a message by its ID
      security:
      - BearerAuth: []
      responses:
        '200':
          description: Message details retrieved successfully
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/MessageResourceEnvelope'
              example:
                code: 200
                message: Message details retrieved
                data:
                  resource:
                    id: '123'
                    content: Hello!
                    sender_id: '1'
                    conversation_id: 'c-1'
                    created_at: '2023-10-01T12:00:00Z'
                    type: text
                    status: delivered
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
      - messages
      operationId: deleteMessage
      summary: Delete a message
      description: Permanently delete a message
      security:
      - BearerAuth: []
      responses:
        '200':
          description: Message deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseSuccessResponse'
              example:
                code: 200
                message: Message deleted successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /messages/{id}/uncomment:
    post:
      tags:
      - messages
      operationId: uncommentMessage
      summary: Remove a comment from a message
      description: Delete a comment from a message
      security:
      - BearerAuth: []
      parameters:
      - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Comment removed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseSuccessResponse'
              example:
                code: 200
                message: Comment removed successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...